// src/components/faculties/types.ts
export type Faculty = {
    id: number;            // internal key
    regNo: string;         // Reg #
    name: string;          // Faculty Name
    pmdcNo: string;        // PMDC #
    doj: string;           // Date of Joining
    department: string;    // Department
    designation: string;   // Designation
    contact: string;       // Contact #
};